avoid double queries at startup
The 10ms is regrettable, but leaving it out didn't work.

The better solution here would be to either ignore starting a concurrent
run of the same query (maybe?) and properly cancel a previous run if we
are starting a new one, since mixing rows from two runs is definitely
wrong.